// src/models/SocietyBuildingUnitTransfersModel.tsx
export interface SocietyBuildingUnitTransfersModel {

    SocietyBuildingUnitTransferId: string;
    SocietyBuildingUnitId: string;
    SocietyMemberId: string;
    TransferDate: Date;
    UEndDate: Date;
    Remarks: string;
    TransferFee: number;
    PaymentDetails: string;
    societySubscriptionId: string;
}